+++ src/AtomUI.Desktop.Controls/Message/WindowMessageManager.cs
@@ -7,7 +7,6 @@ using Avalonia.Controls.Primitives;
 using Avalonia.Interactivity;
 using Avalonia.Layout;
 using Avalonia.Threading;
-using Avalonia.VisualTree;
 
 namespace AtomUI.Desktop.Controls;